Move some lab api classes around
