+++ LabApiUtilities/src/main/com/microsoft/identity/labapi/utilities/exception/LabError.java
@@ -20,17 +20,18 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.labapi.utilities;
+package com.microsoft.identity.labapi.utilities.exception;
 
 /**
- * A factory that can return an implementation of an {@link IJWTParser}.
+ * Represents errors that can occur in a Lab Api flow. This includes flows to authenticate against
+ * the lab api as well as flows to obtain an account via the Lab Api.
  */
-public interface IJWTParserFactory {
-
-    /**
-     * Obtain an instance of an {@link IJWTParser}.
-     *
-     * @return the JWT Parser to use.
-     */
-    IJWTParser getJwtParser();
+public enum LabError {
+    FAILED_TO_GET_TOKEN_FOR_KEYVAULT_USING_CLIENT_SECRET,
+    FAILED_TO_GET_TOKEN_FOR_KEYVAULT_USING_CERTIFICATE,
+    FAILED_TO_GET_SECRET_FROM_KEYVAULT,
+    CERTIFICATE_NOT_FOUND_IN_KEY_STORE,
+    FAILED_TO_GET_ACCOUNT_FROM_LAB,
+    FAILED_TO_GET_SECRET_FROM_LAB,
+    FAILED_TO_CREATE_TEMP_USER;
 }
