+++ LabApiUtilities/src/test/com/microsoft/identity/labapi/utilities/authentication/LabApiAuthenticationClientTest.java
@@ -22,11 +22,10 @@
 // THE SOFTWARE.
 package com.microsoft.identity.labapi.utilities.authentication;
 
-import com.microsoft.identity.labapi.utilities.Constants;
-import com.microsoft.identity.labapi.utilities.IJWTParser;
-import com.microsoft.identity.labapi.utilities.JWTParserFactory;
+import com.microsoft.identity.labapi.utilities.jwt.IJWTParser;
+import com.microsoft.identity.labapi.utilities.jwt.JWTParserFactory;
 import com.microsoft.identity.labapi.utilities.TestBuildConfig;
-import com.microsoft.identity.labapi.utilities.authentication.exception.LabApiException;
+import com.microsoft.identity.labapi.utilities.exception.LabApiException;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -50,8 +49,8 @@ public class LabApiAuthenticationClientTest {
             Assert.assertNotNull(accessToken);
             Assert.assertNotEquals("", accessToken);
             Assert.assertEquals(
-                    Constants.LAB_API_TOKEN_AUDIENCE,
-                    ((List<String>) jwtParser.parseJWT(accessToken).get(Constants.AUDIENCE_CLAIM)).get(0)
+                    LabAuthenticationConstants.LAB_API_TOKEN_AUDIENCE,
+                    ((List<String>) jwtParser.parseJWT(accessToken).get(LabAuthenticationConstants.AUDIENCE_CLAIM)).get(0)
             );
         } catch (final LabApiException e) {
             throw new AssertionError(e);
@@ -68,8 +67,8 @@ public class LabApiAuthenticationClientTest {
             Assert.assertNotNull(accessToken);
             Assert.assertNotEquals("", accessToken);
             Assert.assertEquals(
-                    Constants.LAB_API_TOKEN_AUDIENCE,
-                    ((List<String>) jwtParser.parseJWT(accessToken).get(Constants.AUDIENCE_CLAIM)).get(0)
+                    LabAuthenticationConstants.LAB_API_TOKEN_AUDIENCE,
+                    ((List<String>) jwtParser.parseJWT(accessToken).get(LabAuthenticationConstants.AUDIENCE_CLAIM)).get(0)
             );
         } catch (final LabApiException e) {
             throw new AssertionError(e);
