+++ LabApiUtilities/src/test/com/microsoft/identity/labapi/utilties/authentication/common/CertificateCredentialTest.java
@@ -25,8 +25,8 @@ package com.microsoft.identity.labapi.utilties.authentication.common;
 import com.microsoft.identity.labapi.utilities.authentication.common.CertificateCredential;
 import com.microsoft.identity.labapi.utilities.authentication.common.ClientCertificateMetadata;
 import com.microsoft.identity.labapi.utilities.authentication.common.KeyStoreConfiguration;
-import com.microsoft.identity.labapi.utilities.authentication.exception.LabApiException;
-import com.microsoft.identity.labapi.utilities.authentication.exception.LabError;
+import com.microsoft.identity.labapi.utilities.exception.LabApiException;
+import com.microsoft.identity.labapi.utilities.exception.LabError;
 
 import org.junit.Assert;
 import org.junit.Test;
